fix: full table lookup with gorm
With an empty slice, `db.Find(..., empty)`ends up a full table lookup,
instead of returning nothing.

A `Limit` clause should always be added to avoid possible misuses.

+++ gateway/biz/handler/douyin/core/feed_service.go
@@ -38,7 +38,7 @@ func Feed(ctx context.Context, c *app.RequestContext) {
 
 	r, err := services.Feed.Feed(ctx, &rpc.DouyinFeedRequest{
 		LatestTime:    req.LatestTime,
-		RequestUserId: int64(id),
+		RequestUserId: id,
 	})
 	if err != nil {
 		utils.ErrorRpcTimeout.Write(c)
